+++ CHANGELOG.md
@@ -6,7 +6,11 @@
 
 ### Bug Fixes
 
-- TODO
+- Fix `NameError: undefined local variable or method `config` for Rails RSpec
+  configuration (Aaron Kromer, #1)
+- Fix model factory build issue in which registered template attributes, which
+  use symbol keys, are not replaced by custom attributes using string keys
+  (Aaron Kromer, #1)
 
 
 ## 0.1.0 (March 14, 2018)